+++ src/api/open_source/open_source_translate_api.py
@@ -1,20 +1,15 @@
+from typing import List, Dict
+
 import pandas as pd
 from langchain_core.documents import Document
 from langchain_text_splitters import RecursiveCharacterTextSplitter
 
 from src.api.i_translate_api import ITranslateAPI
-from src.client.opus_client import OpusClient
-from typing import Dict, List
 
 
-class OpusTranslateAPI(ITranslateAPI):
-
-    def __init__(self, from_language: str, to_languages: List[str]):
+class OpenSourceTranslateAPI(ITranslateAPI):
+    def __init__(self, from_language, to_languages):
         super().__init__(from_language, to_languages)
-        self.models: Dict[str, OpusClient] = {}
-        for to_language in to_languages:
-            self.models[to_language] = OpusClient(from_language=from_language, to_language=to_language)
-
         self.text_splitter = RecursiveCharacterTextSplitter(
             chunk_size=512,
             chunk_overlap=20,
@@ -36,12 +31,7 @@ class OpusTranslateAPI(ITranslateAPI):
         return result
 
     def _translate(self, text: str, to_language: str) -> str:
-        result = ""
-        chunks = self._get_chunks(text)
-        for chunk in chunks:
-            translated_chunk = self.models[to_language].translate(chunk)
-            result += translated_chunk
-        return result
+        pass
 
     def _get_chunks(self, text) -> List[str]:
         chunks: List[Document] = self.text_splitter.create_documents([text])
