+++ src/client/opus_client.py
@@ -7,7 +7,6 @@ class OpusClient(object):
         model_name = f"Helsinki-NLP/opus-mt-{from_language}-{to_language}"
         self.tokenizer = AutoTokenizer.from_pretrained(model_name)
         self.model = AutoModelForSeq2SeqLM.from_pretrained(model_name)
-        self.to_language = to_language
 
     def translate(self, text) -> str:
         encoded_input = self.tokenizer(text, return_tensors="pt", padding=True)
